+++ DefaultSDL/Rect.h
@@ -5,8 +5,8 @@
 class Rect
 {
 private:
-	int width, height;
-	int xCoord, yCoord;
+	float width, height;
+	float xCoord, yCoord;
 	int red, green, blue, alpha;
 	SDL_Texture* texture;
 
@@ -14,16 +14,16 @@ private:
 
 public:
 
-	Rect(int w, int h, int x, int y, int r, int g, int b, int a);
-	Rect(int w, int h, int x, int y, const std::string& image_path);
+	Rect(float w, float h, float x, float y, int r, int g, int b, int a);
+	Rect(float w, float h, float x, float y, const std::string& image_path);
 	~Rect();
 
 	void draw()const;
 	void pollEvents(SDL_Event& event);
 
-	void setX(int x) { xCoord = x; }
-	void setY(int y) { yCoord = y; }
+	void setX(float x) { xCoord = x; }
+	void setY(float y) { yCoord = y; }
 
-	void moveX(int x) { xCoord += x; }
-	void moveY(int y) { yCoord += y; }
+	void moveX(float x) { xCoord += x; }
+	void moveY(float y) { yCoord += y; }
 };
